one failing test, most issued resolved
