+++ tests/functional/sources/test_source_freshness.py
@@ -1,302 +1,375 @@
-import json
 import os
-from datetime import datetime, timedelta
+import json
 import pytest
-import yaml
+from datetime import datetime, timedelta
 
-from dbt.exceptions import ParsingException
-import dbt.tracking
 import dbt.version
+from tests.functional.sources.common_source_setup import BaseSourcesTest
+from tests.functional.sources.fixtures import (
+    error_models__schema_yml,
+    error_models__model_sql,
+    filtered_models__schema_yml,
+    override_freshness_models__schema_yml,
+)
 
-from dbt.tests.util import run_dbt
-from tests.functional.sources.fixtures import error_models,override_freshness_models,models,malformed_models,filtered_models,macros,seeds,malformed_schema_tests,project_files # noqa: F401
 
-# from test.integration.base import DBTIntegrationTest, use_profile, AnyFloat, \
-#     AnyStringWith
+# put these here for now to test
+class AnyStringWith:
+    def __init__(self, contains=None):
+        self.contains = contains
+
+    def __eq__(self, other):
+        if not isinstance(other, str):
+            return False
+
+        if self.contains is None:
+            return True
+
+        return self.contains in other
+
+    def __repr__(self):
+        return "AnyStringWith<{!r}>".format(self.contains)
 
 
-    def _set_updated_at_to(self, delta):
+class AnyFloat:
+    """Any float. Use this in assertEqual() calls to assert that it is a float."""
+
+    def __eq__(self, other):
+        return isinstance(other, float)
+
+
+class SuccessfulSourceFreshnessTest(BaseSourcesTest):
+    @pytest.fixture(scope="class", autouse=True)
+    def setUp(self, project):
+        self.run_dbt_with_vars(project, ["seed"])
+        pytest._id = 101
+        pytest.freshness_start_time = datetime.utcnow()
+        # this is the db initial value
+        pytest.last_inserted_time = "2016-09-19T14:45:51+00:00"
+
+        os.environ["DBT_ENV_CUSTOM_ENV_key"] = "value"
+
+        yield
+
+        del os.environ["DBT_ENV_CUSTOM_ENV_key"]
+
+    def _set_updated_at_to(self, project, delta):
         insert_time = datetime.utcnow() + delta
         timestr = insert_time.strftime("%Y-%m-%d %H:%M:%S")
         # favorite_color,id,first_name,email,ip_address,updated_at
-        insert_id = self._id
-        self._id += 1
+        insert_id = pytest._id
+        pytest._id += 1
+        quoted_columns = ",".join(
+            project.adapter.quote(c)
+            for c in ("favorite_color", "id", "first_name", "email", "ip_address", "updated_at")
+        )
+        kwargs = {
+            "schema": project.test_schema,
+            "time": timestr,
+            "id": insert_id,
+            "source": project.adapter.quote("source"),
+            "quoted_columns": quoted_columns,
+        }
         raw_sql = """INSERT INTO {schema}.{source}
             ({quoted_columns})
         VALUES (
             'blue',{id},'Jake','abc@example.com','192.168.1.1','{time}'
-        )"""
-        quoted_columns = ','.join(
-            self.adapter.quote(c) for c in
-            ('favorite_color', 'id', 'first_name',
-             'email', 'ip_address', 'updated_at')
+        )""".format(
+            **kwargs
         )
-        self.run_sql(
-            raw_sql,
-            kwargs={
-                'schema': self.unique_schema(),
-                'time': timestr,
-                'id': insert_id,
-                'source': self.adapter.quote('source'),
-                'quoted_columns': quoted_columns,
-            }
-        )
-        self.last_inserted_time = insert_time.strftime(
-            "%Y-%m-%dT%H:%M:%S+00:00")
+        project.run_sql(raw_sql)
+        pytest.last_inserted_time = insert_time.strftime("%Y-%m-%dT%H:%M:%S+00:00")
 
+    def assertBetween(self, timestr, start, end=None):
+        datefmt = "%Y-%m-%dT%H:%M:%S.%fZ"
+        if end is None:
+            end = datetime.utcnow()
 
+        parsed = datetime.strptime(timestr, datefmt)
+
+        assert start <= parsed
+        assert end >= parsed
 
     def _assert_freshness_results(self, path, state):
         assert os.path.exists(path)
         with open(path) as fp:
             data = json.load(fp)
 
-        assert set(data) == {'metadata', 'results', 'elapsed_time'}
-        assert 'generated_at' in data['metadata']
-        assert isinstance(data['elapsed_time'], float)
-        self.assertBetween(data['metadata']['generated_at'],
-                           self.freshness_start_time)
-        assert data['metadata']['dbt_schema_version'] == 'https://schemas.getdbt.com/dbt/sources/v3.json'
-        assert data['metadata']['dbt_version'] == dbt.version.__version__
-        assert data['metadata']['invocation_id'] == dbt.tracking.active_user.invocation_id
-        key = 'key'
-        if os.name == 'nt':
+        assert set(data) == {"metadata", "results", "elapsed_time"}
+        assert "generated_at" in data["metadata"]
+        assert isinstance(data["elapsed_time"], float)
+        self.assertBetween(data["metadata"]["generated_at"], pytest.freshness_start_time)
+        assert (
+            data["metadata"]["dbt_schema_version"]
+            == "https://schemas.getdbt.com/dbt/sources/v3.json"
+        )
+        assert data["metadata"]["dbt_version"] == dbt.version.__version__
+        assert data["metadata"]["invocation_id"] == dbt.tracking.active_user.invocation_id
+        key = "key"
+        if os.name == "nt":
             key = key.upper()
-        assert data['metadata']['env'] == {key: 'value'}
+        assert data["metadata"]["env"] == {key: "value"}
 
-        last_inserted_time = self.last_inserted_time
+        last_inserted_time = pytest.last_inserted_time
 
-        assert len(data['results']) == 1
+        assert len(data["results"]) == 1
 
-        assert data['results'] == [
+        # TODO: replace below calls - could they be mroe sane?
+        assert data["results"] == [
             {
-                'unique_id': 'source.test.test_source.test_table',
-                'max_loaded_at': last_inserted_time,
-                'snapshotted_at': AnyStringWith(),
-                'max_loaded_at_time_ago_in_s': AnyFloat(),
-                'status': state,
-                'criteria': {
-                    'filter': None,
-                    'warn_after': {'count': 10, 'period': 'hour'},
-                    'error_after': {'count': 18, 'period': 'hour'},
+                "unique_id": "source.test.test_source.test_table",
+                "max_loaded_at": last_inserted_time,
+                "snapshotted_at": AnyStringWith(),
+                "max_loaded_at_time_ago_in_s": AnyFloat(),
+                "status": state,
+                "criteria": {
+                    "filter": None,
+                    "warn_after": {"count": 10, "period": "hour"},
+                    "error_after": {"count": 18, "period": "hour"},
                 },
-                'adapter_response': {},
-                'thread_id': AnyStringWith('Thread-'),
-                'execution_time': AnyFloat(),
-                'timing': [
+                "adapter_response": {},
+                "thread_id": AnyStringWith("Thread-"),
+                "execution_time": AnyFloat(),
+                "timing": [
                     {
-                        'name': 'compile',
-                        'started_at': AnyStringWith(),
-                        'completed_at': AnyStringWith(),
+                        "name": "compile",
+                        "started_at": AnyStringWith(),
+                        "completed_at": AnyStringWith(),
                     },
                     {
-                        'name': 'execute',
-                        'started_at': AnyStringWith(),
-                        'completed_at': AnyStringWith(),
-                    }
-                ]
+                        "name": "execute",
+                        "started_at": AnyStringWith(),
+                        "completed_at": AnyStringWith(),
+                    },
+                ],
             }
         ]
 
 
+class TestSourceFreshness(SuccessfulSourceFreshnessTest):
     def test_source_freshness(self, project):
         # test_source.test_table should have a loaded_at field of `updated_at`
         # and a freshness of warn_after: 10 hours, error_after: 18 hours
         # by default, our data set is way out of date!
-        self.freshness_start_time = datetime.utcnow()
+
         results = self.run_dbt_with_vars(
-            ['source', 'freshness', '-o', 'target/error_source.json'],
-            expect_pass=False
+            project, ["source", "freshness", "-o", "target/error_source.json"], expect_pass=False
         )
         assert len(results) == 1
-        assert results[0].status == 'error'
-        self._assert_freshness_results('target/error_source.json', 'error')
+        assert results[0].status == "error"
+        self._assert_freshness_results("target/error_source.json", "error")
 
-        self._set_updated_at_to(timedelta(hours=-12))
-        self.freshness_start_time = datetime.utcnow()
+        self._set_updated_at_to(project, timedelta(hours=-12))
         results = self.run_dbt_with_vars(
-            ['source', 'freshness', '-o', 'target/warn_source.json'],
+            project,
+            ["source", "freshness", "-o", "target/warn_source.json"],
         )
         assert len(results) == 1
-        assert results[0].status == 'warn'
-        self._assert_freshness_results('target/warn_source.json', 'warn')
+        assert results[0].status == "warn"
+        self._assert_freshness_results("target/warn_source.json", "warn")
 
-        self._set_updated_at_to(timedelta(hours=-2))
-        self.freshness_start_time = datetime.utcnow()
+        self._set_updated_at_to(project, timedelta(hours=-2))
         results = self.run_dbt_with_vars(
-            ['source', 'freshness', '-o', 'target/pass_source.json'],
+            project,
+            ["source", "freshness", "-o", "target/pass_source.json"],
         )
         assert len(results) == 1
-        assert results[0].status == 'pass'
-        self._assert_freshness_results('target/pass_source.json', 'pass')
+        assert results[0].status == "pass"
+        self._assert_freshness_results("target/pass_source.json", "pass")
 
 
-    def test_source_snapshot_freshness(self, project, ):
+class TestSourceSnapshotFreshness(SuccessfulSourceFreshnessTest):
+    def test_source_snapshot_freshness(self, project):
         """Ensures that the deprecated command `source snapshot-freshness`
         aliases to `source freshness` command.
         """
-        self.freshness_start_time = datetime.utcnow()
         results = self.run_dbt_with_vars(
-            ['source', 'snapshot-freshness', '-o', 'target/error_source.json'],
-            expect_pass=False
+            project,
+            ["source", "snapshot-freshness", "-o", "target/error_source.json"],
+            expect_pass=False,
         )
         assert len(results) == 1
-        assert results[0].status == 'error'
-        self._assert_freshness_results('target/error_source.json', 'error')
+        assert results[0].status == "error"
+        self._assert_freshness_results("target/error_source.json", "error")
 
-        self._set_updated_at_to(timedelta(hours=-12))
-        self.freshness_start_time = datetime.utcnow()
+        self._set_updated_at_to(project, timedelta(hours=-12))
         results = self.run_dbt_with_vars(
-            ['source', 'snapshot-freshness', '-o', 'target/warn_source.json'],
+            project,
+            ["source", "snapshot-freshness", "-o", "target/warn_source.json"],
         )
         assert len(results) == 1
-        assert results[0].status == 'warn'
-        self._assert_freshness_results('target/warn_source.json', 'warn')
+        assert results[0].status == "warn"
+        self._assert_freshness_results("target/warn_source.json", "warn")
 
-        self._set_updated_at_to(timedelta(hours=-2))
-        self.freshness_start_time = datetime.utcnow()
+        self._set_updated_at_to(project, timedelta(hours=-2))
         results = self.run_dbt_with_vars(
-            ['source', 'snapshot-freshness', '-o', 'target/pass_source.json'],
+            project,
+            ["source", "snapshot-freshness", "-o", "target/pass_source.json"],
         )
         assert len(results) == 1
-        assert results[0].status == 'pass'
-        self._assert_freshness_results('target/pass_source.json', 'pass')
+        assert results[0].status == "pass"
+        self._assert_freshness_results("target/pass_source.json", "pass")
 
-    def test_source_freshness_selection_select(self, project, ):
+
+class TestSourceFreshnessSelection(SuccessfulSourceFreshnessTest):
+    def test_source_freshness_selection_select(self, project):
         """Tests node selection using the --select argument."""
-        self._set_updated_at_to(timedelta(hours=-2))
-        self.freshness_start_time = datetime.utcnow()
+        self._set_updated_at_to(project, timedelta(hours=-2))
         # select source directly
         results = self.run_dbt_with_vars(
-            ['source', 'freshness', '--select',
-                'source:test_source.test_table', '-o', 'target/pass_source.json'],
+            project,
+            [
+                "source",
+                "freshness",
+                "--select",
+                "source:test_source.test_table",
+                "-o",
+                "target/pass_source.json",
+            ],
         )
         assert len(results) == 1
-        assert results[0].status == 'pass'
-        self._assert_freshness_results('target/pass_source.json', 'pass')
+        assert results[0].status == "pass"
+        self._assert_freshness_results("target/pass_source.json", "pass")
 
-    def test_source_freshness_selection_exclude(self, project, ):
+
+class TestSourceFreshnessExclude(SuccessfulSourceFreshnessTest):
+    def test_source_freshness_selection_exclude(self, project):
         """Tests node selection using the --select argument. It 'excludes' the
         only source in the project so it should return no results."""
-        self._set_updated_at_to(timedelta(hours=-2))
-        self.freshness_start_time = datetime.utcnow()
+        self._set_updated_at_to(project, timedelta(hours=-2))
         # exclude source directly
         results = self.run_dbt_with_vars(
-            ['source', 'freshness', '--exclude',
-                'source:test_source.test_table', '-o', 'target/exclude_source.json'],
+            project,
+            [
+                "source",
+                "freshness",
+                "--exclude",
+                "source:test_source.test_table",
+                "-o",
+                "target/exclude_source.json",
+            ],
         )
         assert len(results) == 0
 
-    def test_source_freshness_selection_graph_operation(self, project, ):
+
+class TestSourceFreshnessGraph(SuccessfulSourceFreshnessTest):
+    def test_source_freshness_selection_graph_operation(self, project):
         """Tests node selection using the --select argument with graph
         operations. `+descendant_model` == select all nodes `descendant_model`
         depends on.
         """
-        self._set_updated_at_to(timedelta(hours=-2))
-        self.freshness_start_time = datetime.utcnow()
+        self._set_updated_at_to(project, timedelta(hours=-2))
         # select model ancestors
         results = self.run_dbt_with_vars(
-            ['source', 'freshness', '--select',
-                '+descendant_model', '-o', 'target/ancestor_source.json']
+            project,
+            [
+                "source",
+                "freshness",
+                "--select",
+                "+descendant_model",
+                "-o",
+                "target/ancestor_source.json",
+            ],
         )
         assert len(results) == 1
-        assert results[0].status == 'pass'
-        self._assert_freshness_results('target/ancestor_source.json', 'pass')
+        assert results[0].status == "pass"
+        self._assert_freshness_results("target/ancestor_source.json", "pass")
 
 
+class TestSourceFreshnessErrors(SuccessfulSourceFreshnessTest):
     @pytest.fixture(scope="class")
-    def model_path(self):
-        return "override_freshness_models"
+    def models(self):
+        return {
+            "schema.yml": error_models__schema_yml,
+            "model.sql": error_models__model_sql,
+        }
 
-    def get_result_from_unique_id(data, unique_id):
+    def test_source_freshness_error(self, project):
+        results = self.run_dbt_with_vars(project, ["source", "freshness"], expect_pass=False)
+        assert len(results) == 1
+        assert results[0].status == "runtime error"
+
+
+class TestSourceFreshnessFilter(SuccessfulSourceFreshnessTest):
+    @pytest.fixture(scope="class")
+    def filtered_models(self):
+        return {"schema.yml": filtered_models__schema_yml}
+
+    def test_source_freshness_all_records(self, project):
+        # all records are filtered out
+        self.run_dbt_with_vars(project, ["source", "freshness"], expect_pass=False)
+        # we should insert a record with _id=101 that's fresh, but will still fail
+        # because the filter excludes it
+        self._set_updated_at_to(project, timedelta(hours=-2))
+        self.run_dbt_with_vars(project, ["source", "freshness"], expect_pass=False)
+
+        # we should now insert a record with _id=102 that's fresh, and the filter
+        # includes it
+        self._set_updated_at_to(project, timedelta(hours=-2))
+        self.run_dbt_with_vars(project, ["source", "freshness"], expect_pass=True)
+
+
+class TestOverrideSourceFreshness(SuccessfulSourceFreshnessTest):
+    @pytest.fixture(scope="class")
+    def filtered_models(self):
+        return {"schema.yml": override_freshness_models__schema_yml}
+
+    @staticmethod
+    def get_result_from_unique_id(self, data, unique_id):
         try:
-            return list(filter(lambda x : x['unique_id'] == unique_id,  data['results']))[0]
+            return list(filter(lambda x: x["unique_id"] == unique_id, data["results"]))[0]
         except IndexError:
             raise f"No result for the given unique_id. unique_id={unique_id}"
 
-    def _run_override_source_freshness(self):
-        self._set_updated_at_to(timedelta(hours=-30))
-        self.freshness_start_time = datetime.utcnow()
+    def test_override_source_freshness(self, project):
+        self._set_updated_at_to(project, timedelta(hours=-30))
 
-        path = 'target/pass_source.json'
+        path = "target/pass_source.json"
         results = self.run_dbt_with_vars(
-            ['source', 'freshness', '-o', path],
-            expect_pass=False
+            project, ["source", "freshness", "-o", path], expect_pass=False
         )
-        assert len(results) == 4 # freshness disabled for source_e
+        assert len(results) == 4  # freshness disabled for source_e
 
         assert os.path.exists(path)
         with open(path) as fp:
             data = json.load(fp)
 
-        result_source_a = self.get_result_from_unique_id(data, 'source.test.test_source.source_a')
-        assert result_source_a['status'] == 'error'
-        assert result_source_a['criteria'] == \
-            {
-                'warn_after': {'count': 6, 'period': 'hour'},
-                'error_after': {'count': 24, 'period': 'hour'},
-                'filter': None
-            }
+        result_source_a = self.get_result_from_unique_id(data, "source.test.test_source.source_a")
+        assert result_source_a["status"] == "error"
 
-        result_source_b = self.get_result_from_unique_id(data, 'source.test.test_source.source_b')
-        assert result_source_b['status'] == 'error'
-        assert result_source_b['criteria'] == \
-            {
-                'warn_after': {'count': 6, 'period': 'hour'},
-                'error_after': {'count': 24, 'period': 'hour'},
-                'filter': None
-            }
+        expected = {
+            "warn_after": {"count": 6, "period": "hour"},
+            "error_after": {"count": 24, "period": "hour"},
+            "filter": None,
+        }
+        assert result_source_a["criteria"] == expected
 
-        result_source_c = self.get_result_from_unique_id(data, 'source.test.test_source.source_c')
-        assert result_source_c['status'] == 'warn'
-        assert result_source_c['criteria'] == \
-            {
-                'warn_after': {'count': 6, 'period': 'hour'},
-                'error_after': None,
-                'filter': None
-            }
+        result_source_b = self.get_result_from_unique_id(data, "source.test.test_source.source_b")
+        assert result_source_b["status"] == "error"
 
-        result_source_d = self.get_result_from_unique_id(data, 'source.test.test_source.source_d')
-        assert result_source_d['status'] == 'warn'
-        assert result_source_d['criteria'] == \
-            {
-                'warn_after': {'count': 6, 'period': 'hour'},
-                'error_after': {'count': 72, 'period': 'hour'},
-                'filter': None
-            }
+        expected = {
+            "warn_after": {"count": 6, "period": "hour"},
+            "error_after": {"count": 24, "period": "hour"},
+            "filter": None,
+        }
+        assert result_source_b["criteria"] == expected
 
-    def test_override_source_freshness(self, project, ):
-        self._run_override_source_freshness()
+        result_source_c = self.get_result_from_unique_id(data, "source.test.test_source.source_c")
+        assert result_source_c["status"] == "warn"
 
-    @pytest.fixture(scope="class")
-    def model_path(self):
-        return "error_models"
+        expected = {
+            "warn_after": {"count": 6, "period": "hour"},
+            "error_after": None,
+            "filter": None,
+        }
+        assert result_source_c["criteria"] == expected
 
-    def test_error(self, project, ):
-        results = self.run_dbt_with_vars(
-            ['source', 'freshness'],
-            expect_pass=False
-        )
-        assert len(results) == 1
-        assert results[0].status == 'runtime error'
+        result_source_d = self.get_result_from_unique_id(data, "source.test.test_source.source_d")
+        assert result_source_d["status"] == "warn"
 
-
-    @pytest.fixture(scope="class")
-    def model_path(self):
-        return 'filtered_models'
-
-    def test_all_records(self, project, ):
-        # all records are filtered out
-        self.run_dbt_with_vars(
-            ['source', 'freshness'], expect_pass=False)
-        # we should insert a record with #101 that's fresh, but will still fail
-        # because the filter excludes it
-        self._set_updated_at_to(timedelta(hours=-2))
-        self.run_dbt_with_vars(
-            ['source', 'freshness'], expect_pass=False)
-
-        # we should now insert a record with #102 that's fresh, and the filter
-        # includes it
-        self._set_updated_at_to(timedelta(hours=-2))
-        results = self.run_dbt_with_vars(
-            ['source', 'freshness'], expect_pass=True)
+        expected = {
+            "warn_after": {"count": 6, "period": "hour"},
+            "error_after": {"count": 72, "period": "hour"},
+            "filter": None,
+        }
+        assert result_source_d["criteria"] == expected
